Исправления наставника 1
